# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * TradeVision AI Pro API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type AnalysisStatus = typeof AnalysisStatus[keyof typeof AnalysisStatus];


export const AnalysisStatus = {
  pending: 'pending',
  complete: 'complete',
  error: 'error',
} as const;

export interface Analysis {
  id: number;
  /** @nullable */
  symbol?: string | null;
  /** @nullable */
  timeframe?: string | null;
  /** @nullable */
  chartType?: string | null;
  /** @nullable */
  imageBase64?: string | null;
  status: AnalysisStatus;
  /** @nullable */
  trend?: string | null;
  /** @nullable */
  trendStrength?: number | null;
  /** @nullable */
  patterns?: string | null;
  /** @nullable */
  indicators?: string | null;
  /** @nullable */
  supportLevels?: string | null;
  /** @nullable */
  resistanceLevels?: string | null;
  /** @nullable */
  tradeDirection?: string | null;
  /** @nullable */
  entryPrice?: string | null;
  /** @nullable */
  stopLoss?: string | null;
  /** @nullable */
  takeProfit1?: string | null;
  /** @nullable */
  takeProfit2?: string | null;
  /** @nullable */
  takeProfit3?: string | null;
  /** @nullable */
  riskReward?: string | null;
  /** @nullable */
  confidence?: number | null;
  /** @nullable */
  confidenceLabel?: string | null;
  /** @nullable */
  explanation?: string | null;
  /** @nullable */
  strengths?: string | null;
  /** @nullable */
  weaknesses?: string | null;
  /** @nullable */
  risks?: string | null;
  /** @nullable */
  invalidationLevel?: string | null;
  createdAt: string;
}

export interface AnalysisInput {
  symbol?: string;
  timeframe?: string;
  imageBase64: string;
}

export type AnalysisStatsTopPatternsItem = {
  pattern: string;
  count: number;
};

export interface AnalysisStats {
  totalAnalyses: number;
  bullishCount: number;
  bearishCount: number;
  neutralCount: number;
  avgConfidence: number;
  topPatterns: AnalysisStatsTopPatternsItem[];
}

export type WatchlistItemCategory = typeof WatchlistItemCategory[keyof typeof WatchlistItemCategory];


export const WatchlistItemCategory = {
  crypto: 'crypto',
  forex: 'forex',
  stocks: 'stocks',
  commodities: 'commodities',
  indices: 'indices',
} as const;

export interface WatchlistItem {
  id: number;
  symbol: string;
  /** @nullable */
  name?: string | null;
  category: WatchlistItemCategory;
  /** @nullable */
  notes?: string | null;
  addedAt: string;
}

export type WatchlistItemInputCategory = typeof WatchlistItemInputCategory[keyof typeof WatchlistItemInputCategory];


export const WatchlistItemInputCategory = {
  crypto: 'crypto',
  forex: 'forex',
  stocks: 'stocks',
  commodities: 'commodities',
  indices: 'indices',
} as const;

export interface WatchlistItemInput {
  symbol: string;
  name?: string;
  category: WatchlistItemInputCategory;
  notes?: string;
}

export type JournalEntryDirection = typeof JournalEntryDirection[keyof typeof JournalEntryDirection];


export const JournalEntryDirection = {
  long: 'long',
  short: 'short',
} as const;

/**
 * @nullable
 */
export type JournalEntryOutcome = typeof JournalEntryOutcome[keyof typeof JournalEntryOutcome] | null;


export const JournalEntryOutcome = {
  win: 'win',
  loss: 'loss',
  breakeven: 'breakeven',
} as const;

export interface JournalEntry {
  id: number;
  symbol: string;
  direction: JournalEntryDirection;
  entryPrice: number;
  /** @nullable */
  exitPrice?: number | null;
  /** @nullable */
  stopLoss?: number | null;
  /** @nullable */
  takeProfit?: number | null;
  /** @nullable */
  positionSize?: number | null;
  /** @nullable */
  pnl?: number | null;
  /** @nullable */
  riskReward?: number | null;
  /** @nullable */
  outcome?: JournalEntryOutcome;
  /** @nullable */
  strategy?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  chartImageBase64?: string | null;
  /** @nullable */
  analysisId?: number | null;
  tradeDate: string;
  createdAt: string;
}

export type JournalEntryInputDirection = typeof JournalEntryInputDirection[keyof typeof JournalEntryInputDirection];


export const JournalEntryInputDirection = {
  long: 'long',
  short: 'short',
} as const;

export type JournalEntryInputOutcome = typeof JournalEntryInputOutcome[keyof typeof JournalEntryInputOutcome];


export const JournalEntryInputOutcome = {
  win: 'win',
  loss: 'loss',
  breakeven: 'breakeven',
} as const;

export interface JournalEntryInput {
  symbol: string;
  direction: JournalEntryInputDirection;
  entryPrice: number;
  exitPrice?: number;
  stopLoss?: number;
  takeProfit?: number;
  positionSize?: number;
  pnl?: number;
  riskReward?: number;
  outcome?: JournalEntryInputOutcome;
  strategy?: string;
  notes?: string;
  chartImageBase64?: string;
  analysisId?: number;
  tradeDate?: string;
}

export type JournalEntryUpdateOutcome = typeof JournalEntryUpdateOutcome[keyof typeof JournalEntryUpdateOutcome];


export const JournalEntryUpdateOutcome = {
  win: 'win',
  loss: 'loss',
  breakeven: 'breakeven',
} as const;

export interface JournalEntryUpdate {
  exitPrice?: number;
  stopLoss?: number;
  takeProfit?: number;
  positionSize?: number;
  pnl?: number;
  riskReward?: number;
  outcome?: JournalEntryUpdateOutcome;
  strategy?: string;
  notes?: string;
}

export interface JournalStats {
  totalTrades: number;
  winCount: number;
  lossCount: number;
  breakevenCount: number;
  winRate: number;
  avgRiskReward: number;
  totalPnl: number;
  /** @nullable */
  bestTrade: number | null;
  /** @nullable */
  worstTrade: number | null;
  avgPnl: number;
  /** @nullable */
  sharpeRatio: number | null;
}

export interface PortfolioSummary {
  totalTrades: number;
  winRate: number;
  totalPnl: number;
  avgRiskReward: number;
  /** @nullable */
  sharpeRatio: number | null;
  totalAnalyses: number;
  watchlistCount: number;
}

export interface EquityPoint {
  date: string;
  equity: number;
  pnl: number;
}

export interface MonthlyPerformance {
  month: string;
  pnl: number;
  trades: number;
  winRate: number;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}