# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TradeVision AI Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analyses:
    get:
      operationId: listAnalyses
      tags: [analyses]
      summary: List all chart analyses
      responses:
        "200":
          description: List of analyses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Analysis"
    post:
      operationId: createAnalysis
      tags: [analyses]
      summary: Upload a chart image and trigger AI analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalysisInput"
      responses:
        "201":
          description: Analysis created (may be pending while AI processes)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
...
  /analyses/{id}:
    get:
      operationId: getAnalysis
      tags: [analyses]
      summary: Get a single analysis by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteAnalysis
      tags: [analyses]
      summary: Delete an analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Watchlist ────────────────────────────────────────────────────────────────
...
  /journal:
    get:
      operationId: listJournalEntries
      tags: [journal]
      summary: List all trade journal entries
      responses:
        "200":
          description: Journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
    post:
      operationId: createJournalEntry
      tags: [journal]
      summary: Create a trade journal entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
...
  /journal/{id}:
    get:
      operationId: getJournalEntry
      tags: [journal]
      summary: Get a single journal entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Journal entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateJournalEntry
      tags: [journal]
      summary: Update a journal entry (e.g. record outcome)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalEntryUpdate"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteJournalEntry
      tags: [journal]
      summary: Delete a journal entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Portfolio Dashboard ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Analysis ──────────────────────────────────────────────────────────────
    Analysis:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: ["string", "null"]
        timeframe:
          type: ["string", "null"]
        chartType:
          type: ["string", "null"]
        imageBase64:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, complete, error]
        # AI result fields
        trend:
          type: ["string", "null"]
        trendStrength:
          type: ["number", "null"]
        patterns:
          type: ["string", "null"]
        indicators:
          type: ["string", "null"]
        supportLevels:
          type: ["string", "null"]
        resistanceLevels:
          type: ["string", "null"]
        tradeDirection:
          type: ["string", "null"]
        entryPrice:
          type: ["string", "null"]
        stopLoss:
          type: ["string", "null"]
        takeProfit1:
          type: ["string", "null"]
        takeProfit2:
          type: ["string", "null"]
        takeProfit3:
          type: ["string", "null"]
        riskReward:
          type: ["string", "null"]
        confidence:
          type: ["number", "null"]
        confidenceLabel:
          type: ["string", "null"]
        explanation:
          type: ["string", "null"]
        strengths:
          type: ["string", "null"]
        weaknesses:
          type: ["string", "null"]
        risks:
          type: ["string", "null"]
        invalidationLevel:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - status
        - createdAt
    AnalysisInput:
      type: object
      properties:
        symbol:
          type: string
        timeframe:
          type: string
        imageBase64:
          type: string
      required:
        - imageBase64
...
    JournalEntry:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        direction:
          type: string
          enum: [long, short]
        entryPrice:
          type: number
        exitPrice:
          type: ["number", "null"]
        stopLoss:
          type: ["number", "null"]
        takeProfit:
          type: ["number", "null"]
        positionSize:
          type: ["number", "null"]
        pnl:
          type: ["number", "null"]
        riskReward:
          type: ["number", "null"]
        outcome:
          type: ["string", "null"]
          enum: [win, loss, breakeven, null]
        strategy:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        chartImageBase64:
          type: ["string", "null"]
        analysisId:
          type: ["integer", "null"]
        tradeDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - symbol
        - direction
        - entryPrice
        - tradeDate
        - createdAt
    JournalEntryInput:
      type: object
      properties:
        symbol:
          type: string
        direction:
          type: string
          enum: [long, short]
        entryPrice:
          type: number
        exitPrice:
          type: number
        stopLoss:
          type: number
        takeProfit:
          type: number
        positionSize:
          type: number
        pnl:
          type: number
        riskReward:
          type: number
        outcome:
          type: string
          enum: [win, loss, breakeven]
        strategy:
          type: string
        notes:
          type: string
        chartImageBase64:
          type: string
        analysisId:
          type: integer
        tradeDate:
          type: string
          format: date-time
      required:
        - symbol
        - direction
        - entryPrice
    JournalEntryUpdate:
      type: object
      properties:
        exitPrice:
          type: number
        stopLoss:
          type: number
        takeProfit:
          type: number
        positionSize:
          type: number
        pnl:
          type: number
        riskReward:
          type: number
        outcome:
          type: string
          enum: [win, loss, breakeven]
        strategy:
          type: string
        notes:
          type: string
...