# lib/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * TradeVision AI Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisStatus } from './analysisStatus';

export interface Analysis {
  id: number;
  /** @nullable */
  symbol?: string | null;
  /** @nullable */
  timeframe?: string | null;
  /** @nullable */
  chartType?: string | null;
  /** @nullable */
  imageBase64?: string | null;
  status: AnalysisStatus;
  /** @nullable */
  trend?: string | null;
  /** @nullable */
  trendStrength?: number | null;
  /** @nullable */
  patterns?: string | null;
  /** @nullable */
  indicators?: string | null;
  /** @nullable */
  supportLevels?: string | null;
  /** @nullable */
  resistanceLevels?: string | null;
  /** @nullable */
  tradeDirection?: string | null;
  /** @nullable */
  entryPrice?: string | null;
  /** @nullable */
  stopLoss?: string | null;
  /** @nullable */
  takeProfit1?: string | null;
  /** @nullable */
  takeProfit2?: string | null;
  /** @nullable */
  takeProfit3?: string | null;
  /** @nullable */
  riskReward?: string | null;
  /** @nullable */
  confidence?: number | null;
  /** @nullable */
  confidenceLabel?: string | null;
  /** @nullable */
  explanation?: string | null;
  /** @nullable */
  strengths?: string | null;
  /** @nullable */
  weaknesses?: string | null;
  /** @nullable */
  risks?: string | null;
  /** @nullable */
  invalidationLevel?: string | null;
  createdAt: Date;
}