# lib/api-zod/src/generated/types/journalEntry.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * TradeVision AI Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { JournalEntryDirection } from './journalEntryDirection';
import type { JournalEntryOutcome } from './journalEntryOutcome';

export interface JournalEntry {
  id: number;
  symbol: string;
  direction: JournalEntryDirection;
  entryPrice: number;
  /** @nullable */
  exitPrice?: number | null;
  /** @nullable */
  stopLoss?: number | null;
  /** @nullable */
  takeProfit?: number | null;
  /** @nullable */
  positionSize?: number | null;
  /** @nullable */
  pnl?: number | null;
  /** @nullable */
  riskReward?: number | null;
  /** @nullable */
  outcome?: JournalEntryOutcome;
  /** @nullable */
  strategy?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  chartImageBase64?: string | null;
  /** @nullable */
  analysisId?: number | null;
  tradeDate: Date;
  createdAt: Date;
}