# lib/api-zod/src/generated/types/journalEntryUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * TradeVision AI Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { JournalEntryUpdateOutcome } from './journalEntryUpdateOutcome';

export interface JournalEntryUpdate {
  exitPrice?: number;
  stopLoss?: number;
  takeProfit?: number;
  positionSize?: number;
  pnl?: number;
  riskReward?: number;
  outcome?: JournalEntryUpdateOutcome;
  strategy?: string;
  notes?: string;
}